refactor: page별로 파일 분리 - IssueDetail - 이슈상세 페이지 파일 분리
