+++ src/pages/IssueDetail/issuedetail.tsx
@@ -1,12 +1,11 @@
-
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import Grid from '@material-ui/core/Grid'
+import Grid from '@material-ui/core/Grid';
 import Container from '@material-ui/core/Container';
-import AppbarShift from './appbarshift';
-import IssueDetailTabs from './issuedetailtap';
-import IssueDetailHeader from './issuedetailheader';
+import AppbarShift from '../layout/appbarshift';
+import IssueDetailTabs from './components/issuedetailtap';
+import IssueDetailHeader from './components/issuedetailheader';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -24,7 +23,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function IssueDetail():JSX.Element {
+export default function IssueDetail(): JSX.Element {
   const classes = useStyles();
   return (
     <div className={classes.root}>
